feat(network.py): improve ping command to use telegram API
The ping command is updated to use the Telegram API's
`raw.functions.Ping` method for more accurate latency
measurement. This replaces the previous method of measuring
latency by calculating the time it takes to send and receive
a message, which was less precise. The new method directly
measures the time it takes for the Telegram server to respond
to a ping request.

Signed-off-by: Troublescope <tomiprs25@gmail.com>

feat(network.py): change ping message to request response time to improve semantics

Signed-off-by: Troublescope <tomiprs25@gmail.com>