feat(telegram_bot.py): add log_chat attribute to TelegramBot class feat(main.py): implement .logchat command to set or get log chat ID feat(notes.py): remove log_chat config and use bot.log_chat instead
The main goal of these changes is to allow users to configure a log chat ID for the bot.
The log chat ID is used to send error messages and other important information about the bot.
The .logchat command allows users to set or get the log chat ID.
The TelegramBot class now has a log_chat attribute that stores the log chat ID.
The Notes module now uses the bot.log_chat attribute instead of the log_chat config.
This makes it easier to configure the log chat ID and ensures that all modules use the same log chat ID.

Signed-off-by: Troublescope <tomiprs25@gmail.com>