feat(pyproject.toml): use custom pyrogram fork and update yt-dlp
This commit changes the pyrogram dependency to a custom fork
from TelegramPlayground and updates yt-dlp to the latest version.
The custom fork is used to address specific issues or add
functionality not yet available in the official pyrogram package.
The yt-dlp update ensures compatibility and access to the latest
features and bug fixes.

Signed-off-by: Troublescope <tomiprs25@gmail.com>